Fixed bug with sessions on pannel
When an user spam F5 key ( or refresh multiple time ) on the pannel, Discord may rate limit access to the user data. In this case, the session is destroyed, and user sent on index page.

If an user session was destroyed while getting his guilds, the undefined session was sent to the pannel, and a crash occured. A check was added, the bug should be fixed.