Edited some slash command blocks
Edited slash_command_creator and slash_command_reply blocks.

We now ask if replies are ephemeral when creating the command, and not at every reply block anymore.
We need to do that since we can't use both ephemeral and normal replies

A new column Ephemeral (boolean) was added to database