Added errors enum in xml_to_js
This functionality should now be ready to be tested.
When an error occur in back-end while working on server workspace, we can tell to user what caused the error

Added enums in Xml_to_JS, send_workspace.js; and server
Added a switch in guild_panel to support every errors sent by back-end.