Added new PLPGSQL function to manage data storages
This function can be called with `SELECT f_update_data_storages_for_guild(guildId, storagesNames...);`.

Storages names must start with `S` or `I`. It will allow us to detect the type of this storage ( String or Int ? )

Any storage name that doesn't exists in database for this server will be created. If the storage exists in database, but isn't in function args, it will be deleted.

Usage example : `SELECT f_update_data_storages_for_guild('893991454798643230', 'Ixp');` will create the storage nammed 'Ixp', that contain Integers. As long this name is in command args, the storage will be saved in database.
From user's point of view, the storage name will be 'xp'.