New table : data_storage_size
We can now get the number of rows stored in a data storage.

Each time an INSERT or DELETE is done in stored_data table, a trigger will update the data_storage_size table.
Rows in data_storage_size are created or deleted in the function f_update_data_storages_for_guild