"Sent message" will now contain sent embeds
Technically, a normal message, and a message containing an embed are nearly the same, so we can easily save sent embeds in this block.

When an embed message is sent, users can now get the sent message with the "sent message" block. It will allow creation of threads on embeds, for example.