Forms must have unique names : Frontend
Due to database and customID limitations, we can't let users using twice the same form name.

When defining the code that will be executed by the bot, we must find a custom ID. This ID must be craftable easily and within the generator, disabling the possibility of just using a commandId or something related to database.
As that was the only values easily available, I decided to use the server id and the command name